feat: move radio-button-field to f36-forms and adapt code and documentation
