+++ packages/components/forms/src/checkbox-field/README.mdx
@@ -2,13 +2,13 @@
 title: 'CheckboxField'
 type: 'component'
 status: 'stable'
-slug: /components/inputs/src/checkbox-field/
-github: 'https://github.com/contentful/forma-36/tree/master/packages/components/inputs/src/checkbox-field'
+slug: /components/forms/src/checkbox-field/
+github: 'https://github.com/contentful/forma-36/tree/master/packages/components/forms/src/checkbox-field'
 storybook: 'https://f36-storybook.contentful.com/?path=/story/components-checkbox-field--overview'
 typescript: ./CheckboxField.tsx
 ---
 
-Use CheckboxField to give users the possibility to select multiple elements, zero, one or more items. This component can be used for example in forms or terms and conditions. It contains a checkbox with its label. For a single choice, please consider using [RadioButtonField component](https://f36.contentful.com/components/radio-button-field/).
+Use CheckboxField to give users the possibility to select multiple elements, zero, one or more items. This component can be used for example in forms or terms and conditions. It contains a checkbox with its label. For a single choice, please consider using [RadioButtonField component](https://f36.contentful.com/components/forms/src/radio-button-field/).
 
 ## Table of contents
 
@@ -34,7 +34,7 @@ NOTE: When attaching a help text to a checkbox, the medium label has to be used 
 ```jsx
 // import {CheckboxField} from '@contentful/f36-forms';
 
-<CheckboxField label="Label text" />
+<CheckboxField id="randomId" label="Label text" />
 ```
 
 - **With helper text** - Helper text will be displayed below the label.
