+++ packages/components/forms/src/controlled-input-field/README.mdx
@@ -2,8 +2,8 @@
 title: 'ControlledInputField'
 type: 'component'
 status: 'stable'
-slug: /components/inputs/src/controlled-input-field/
-github: 'https://github.com/contentful/forma-36/tree/master/packages/components/inputs/src/controlled-input-field'
+slug: /components/forms/src/controlled-input-field/
+github: 'https://github.com/contentful/forma-36/tree/master/packages/components/forms/src/controlled-input-field'
 storybook: 'https://f36-storybook.contentful.com/?path=/story/components-controlled-input-field--overview'
 typescript: ./ControlledInputField.tsx
 ---
@@ -80,7 +80,7 @@ function RadioButtonFieldExample() {
 ## Content recommendations
 
 - Try to use labels with short, scannable text.
-- Prefer using the components CheckboxField or RadioButtonField instead of directly using the ControlledInputField.
+- Prefer using the components [CheckboxField](https://f36.contentful.com/components/forms/src/checkbox-field/) or [RadioButtonField](https://f36.contentful.com/components/forms/src/radio-button-field/) instead of directly using the ControlledInputField.
 
 ## Accessibility
