+++ packages/components/forms/src/index.ts
@@ -1,3 +1,4 @@
 export * from './Label';
 export * from './controlled-input-field';
 export * from './checkbox-field';
+export * from './radio-button-field';
