+++ packages/components/inputs/src/controlled-input/README.mdx
@@ -80,7 +80,7 @@ function RadioButtonExample() {
 ## Content recommendations
 
 - Try to use labels with short, scannable text.
-- Prefer using the components Checkbox or RadioButton instead of directly using the ControlledInput.
+- Prefer using the components [Checkbox](https://f36.contentful.com/components/inputs/src/checkbox/) or [RadioButton](https://f36.contentful.com/components/inputs/src/radio-button/) instead of directly using the ControlledInput.
 
 ## Accessibility
