--- packages/forma-36-react-components/src/components/RadioButtonField/README.mdx
@@ -1,81 +1,0 @@
----
-title: 'RadioButtonField'
-type: 'component'
-status: 'stable'
-slug: /components/radio-button-field/
-github: 'https://github.com/contentful/forma-36/tree/master/packages/forma-36-react-components/src/components/RadioButtonField'
-storybook: 'https://f36-storybook.contentful.com/?path=/story/form-elements-radiobuttonfield--basic'
----
-
-Use RadioButtonField to give users the possibility to select only one option from a number of choices. It contains a radio input with its label. If needed the component can display an error message and hint message. For a multiple-choice, please consider using [ChecboxField component](https://f36.contentful.com/components/checkbox-field/).
-
-## Table of contents
-
-- [Table of contents](#table-of-contents)
-- [How to use RadioButtonField](#how-to-use-radiobuttonfield)
-- [Code examples](#code-examples)
-- [Content recommendations](#content-recommendations)
-- [Accessibility](#accessibility)
-
-## How to use RadioButtonField
-
-- RadioButtonField should be used in a group. Please include at least two or more choices for users
-- Options provided for users should be mutually exclusive.
-- Make sure choices don’t overlap. Avoid patterns like: 1990-2000, 2000-2010.
-- Be explicit about the action that will follow if the radio button is selected
-
-## Code examples
-
-Basic usage
-
-```jsx
-// import {RadioButtonField} from '@contentful/f36-components';
-
-<RadioButtonField label="Label text" id="randomId" />
-```
-
-Example with 2 options with helper text displayed
-
-```jsx
-// import {RadioButtonField} from '@contentful/f36-components';
-
-function RadioButtonFieldExample() {
-  const [activeOption, setActiveOption] = React.useState('yes');
-
-  return (
-    <Flex flexDirection="column">
-      <RadioButtonField
-        name="radio option 1"
-        id="randomId1"
-        value="yes"
-        checked={activeOption === 'yes'}
-        onChange={(e) => {
-          setActiveOption(e.target.value);
-        }}
-        label="Radio button checked"
-        helpText="Helper text displayed below"
-      />
-      <RadioButtonField
-        name="radio option 2"
-        id="randomId2"
-        value="no"
-        isChecked={activeOption === 'no'}
-        onChange={(e) => {
-          setActiveOption(e.target.value);
-        }}
-        label="Second option"
-        helpText="Helper text displayed below"
-      />
-    </Flex>
-  );
-}
-```
-
-## Content recommendations
-
-- Labels should be concise and provide context.
-- Use sentence-style capitalization (only the first word in a phrase and any proper nouns capitalized)
-
-## Accessibility
-
-- Radio buttons should have a clear and concise label.
